add validation in post router

+++ src/features/posts/routers/handlers/deletePostHandler.ts
@@ -3,6 +3,7 @@ import { RequestWithParams } from "../../../../core/types/request-types";
 import { URIParamsPostIdModel } from "../../models/URIParamsPostModel";
 import { HttpStatus } from "../../../../core/types/http-statuses";
 import { postsRepository } from "../../repositories/posts.repository";
+import { createErrorMessages } from "../../../../core/middlewares/input-validtion-result.middleware";
 
 export function deletePostHandler(
   req: RequestWithParams<URIParamsPostIdModel>,
@@ -11,12 +12,12 @@ export function deletePostHandler(
   const id = String(req.params.id);
   const post = postsRepository.findOneById(id);
 
-  // if (!post) {
-  //   res
-  //     .status(HttpStatus.NotFound)
-  //     .send(createErrorMessages([{ field: "id", message: "Post not found" }]));
-  //   return;
-  // }
+  if (!post) {
+    res
+      .status(HttpStatus.NotFound)
+      .send(createErrorMessages([{ field: "id", message: "Post not found" }]));
+    return;
+  }
 
   postsRepository.deleteById(id);
